pyproject.toml
[project]
name = "benchmarks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-generativeai",
    "google-genai",
    "python-dotenv",
    "streamlit",
    "pandas",
    "openai",
    "lmstudio",
]
